+++ src/core/doomseekerfilepaths.h
@@ -39,6 +39,7 @@ class DataPaths;
 class MAIN_EXPORT DoomseekerFilePaths
 {
 	public:
+		static const QString CACERTS_FILENAME;
 		static const QString IP2C_DATABASE_FILENAME;
 		static const QString IP2C_QT_SEARCH_PATH;
 		static const QString TEMP_SERVER_CONFIG_FILENAME;
@@ -48,6 +49,7 @@ class MAIN_EXPORT DoomseekerFilePaths
 
 		static DataPaths* pDataPaths;
 
+		static QString cacerts();
 		static QString ini();
 		static QString ircIni();
 		static QString ip2cDatabase();